refactor: util 함수 이동
